requirements.txt
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
fonttools==4.53.1
html5lib==1.1
markdown2==2.5.0
pdfkit==1.0.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
six==1.16.0
tinycss2==1.3.0
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
